==> 02_Pandas-main/04_practice.ipynb <==
# Pandas 연습 문제 1
---
### 다음의 데이터를 이용하여 실습 문제를 해결하세요.
| Name | Age | City | Salary |
| --- | --- | --- | --- |
| Alice | 25 | New York | 70000 |
| Bob | 30 | Los Angeles | 80000 |
| Charlie | 35 | Chicago | 120000 |
| David | 40 | Houston | 90000 |
| Eve | 28 | San Francisco | 75000 |


**문제1. DataFrame 객체 생성**
- 위의 데이터를 사용하여 DataFrame객체를 생성하세요.

import pandas as pd

# 예시 데이터 생성
data = {
    'Name': ['Alice', 'Bob', 'Charlie', 'David', 'Eve'],
    'Age': [25, 30, 35, 40, 28],
    'City': ['New York', 'Los Angeles', 'Chicago', 'Houston', 'San Francisco'],
    'Salary': [70000, 80000, 120000, 90000, 75000]
}

df = pd.DataFrame(data)
print(df)

**문제2. 열과 행 선택**
- `Name`과 `City` 열을 선택하여 출력하세요.
- 나이가 30 이상인 행을 선택하여 출력하세요.

# Name과 City 열 선택
df[['Name','City']]

# 나이가 30 이상인 행 선택
print(df[df['Age'] >= 30])

**문제 3: 행과 열 추가**

- `Country` 열을 추가하고, 모두 'USA'로 설정하세요.
- 새로운 행을 추가하세요. (Name: Frank, Age: 33, City: Seattle, Salary: 85000)

# Country 열 추가

df['Country'] = ['USA','USA','USA','USA','USA'] #인덱스 갯수에 맞춰서 넣어줘야함 4개!                
df

# 새로운 행 추가
new_row = ['Frank', '33', "Seattle", '85000']

df.loc[5] = new_row
df

**문제 4: 데이터 삭제**

- `Country` 열을 삭제하세요.
- 'David'가 있는 행을 삭제하세요.

# Country 열 삭제

# 'David'가 있는 행 삭제

**문제 5: 데이터 수정**

- `Age`가 35 이상인 사람들의 `Salary`를 5% 증가시키세요.
- 'Alice'의 `City`를 'Boston'으로 수정하세요.

# Age가 35 이상인 사람들의 Salary 5% 증가

# 'Alice'의 City를 'Boston'으로 수정

**문제 6: 데이터 타입 변환**

- `Salary`의 데이터 타입을 float로 변환하세요.
