==> 02_Pandas-main/06_data_Exploring.ipynb <==
# 6. 데이터 탐색하기
---

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

### DataFrame 확인

**describe()**
- 기술 통계 정보를 요약해서 보여주는 기능
- 옵션없이 실행했을때, 기본적으로 계산가능한 컬럼들에 대해 통계를 계산해준다.

> 기본사용시
> - count: NaN (결측치)를 제외한 데이터 개수
> - mean: 평균값
> - std: 표준편차 (데이터가 얼마나 흩어져 있는지)
> - min: 최솟값
> - 25%: 1사분위수 (하위 25% 지점의 값)
> - 50%: 2사분위수 (중앙값, median과 동일)
> - 75%: 3사분위수 (하위 75% 지점의 값)
> - max: 최댓값

# 데이터를 처음 받았을때 각 열이 어떤 값들로 이루어져있는지, 데이터 분포는 어떤지 빠르게 익히는데 도움됨됨
df.describe()

**info()**

- 데이터프레임의 전체적인 구조와 각 컬럼의 정보를 요약해서 보여준다.
- 데이터 분석 시작 전, 결측치(Missing Data) 확인이나 데이터 타입(Dtype) 확인을 위해 많이 사용한다.

df.info()

df.head() # 처음 5개의 row를 가져옴

df.head(7) # 처음 7개의 row를 가져옴

df.tail() # 마지막 5개의 row를 가져옴

df.values # 데이터에 들어있는 값들 확인

df.index # 인덱스 정보 확인

df.columns # 컬럼 정보 확인

df.shape # row, column 갯수 확인

### Series 확인

df['평균구매액(만원)'].describe()

df['평균구매액(만원)'].max()

df['평균구매액(만원)'].nlargest(3) # 평균구매액 높은 순서대로 3명 데이터

df['평균구매액(만원)'].mean() # 평균구매액 평균

df['평균구매액(만원)'].sum() # 평균구매액 합계

df['주요관심사'].count() # 주요관심사 데이터 개수

df['주요관심사'].unique() # 주요관심사 종류 확인
