==> 02_Pandas-main/12_Preprocessing2.ipynb <==
import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

### 데이터 수정

**replace()**
- 데이터 프레임이나 시리즈의 특정 값을 다른 값으로 바꿔주는 기능

df['거주도시'].replace('울산','창원')

df['거주도시'].replace({'울산':'창원', '시흥':'부천'}, inplace=True)
df

거주도시를 ~~시까지 붙이고싶다면?

df['거주도시'] = df['거주도시'] + '시'
df

고객만족도와 재구매의사 점수의 총점을 보고싶다면?

df['총합']  = df['고객만족도(점)'] + df['재구매의사(점)']
df

filter = df['총합'] >= 175

df.loc[filter,'결과'] = 'GOOD'
df.loc[~filter,'결과'] = 'ㅠ'
df

Column 삭제

df.drop(columns=['총합'], inplace=True)

df

Row 삭제

df.drop

박지성의 주요관심사를 SPORTS로 변경해보자

df.loc['C003','주요관심사'] = 'SPORTS'

df

한유진의 거주도시를 부천시로, 주요관심사를 FASHION으로 수정해보자

df.loc['C007',['주요관심사','거주도시']] = ['FASHION','부천시시']
df.tail(3)
