==> 02_Pandas-main/14_grouping.ipynb <==
# 14. 그룹화
---
- 동일한 값을 가진 것들끼리 합쳐서 통계 또는 평균 등의 값을 계산하기 위해 사용

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

**`groupby()` 함수**
- 데이터를 특정 기준에 따라 그룹화하여 집계 연산을 수행한다.
- 이를 통해 데이터의 패턴을 발견하거나 요약 통계를 얻는 데 매우 유용하다.        
- 기본적인 그룹화`groupby()`는 주어진 열을 기준으로 데이터를 그룹화하고, <br>
    그 결과에 대해 다양한 집계 연산을 적용할 수 있다

group = df.groupby('주요관심사').get_group('Electronics')

group

df.groupby('거주도시').size()

#주요관심사가 electronics 인 데이터의 재구매의사점수 평균

df.groupby('주요관심사').get_group('Electronics')['재구매의사(점)'].mean()

**`agg()`**
- 그룹화된 데이터에 대해 원하는 집계 연산을 적용하여 요약 통계를 얻을 수 있다.

데이터 집계함수를 사용해보기 위한 DataFrame 수정

# 첫번쨰 문자가 대문자인지 아닌지 -> 대문자가 아니면 대문자로 변환
def capitalize(interest):
    #notnull() : NaN 인지 아닌지 NaN이면 False 
    if pd.notnull(interest):
         #interest가 NaN이 아닌경우
         return interest.capitalize() #문자열 첫글자만 대문자로 바꿔주는 함수
    
    #nan이면 그대로 return
    return 'Sports'

df['주요관심사'] = df['주요관심사'].apply(capitalize)
df

주요관심사별로 고객만족도와 재구매의사의 최소,최대,총합,평균 계산

aggregated = df.groupby('주요관심사').agg(
    {   
        '재구매의사(점)': ['min', 'max', 'sum', 'mean'],
        '고객만족도(점)': ['min', 'max', 'sum', 'mean']
    }
)
aggregated

실습문제

import pandas as pd

# 예시 상품 판매 데이터
sales_data = {
    'OrderID': [101, 102, 103, 104, 105, 106, 107, 108, 109, 110],
    'Product': ['Laptop', 'Mouse', 'Laptop', 'Monitor', 'Mouse', 'Keyboard', 'Laptop', 'Mouse', 'Monitor', 'Keyboard'],
    'Quantity': [1, 2, 1, 1, 2, 1, 1, 3, 1, 1],
    'Price': [1200, 25, 1200, 300, 25, 100, 1200, 25, 300, 100],
    'CustomerID': [1001, 1002, 1001, 1003, 1004, 1002, 1001, 1003, 1005, 1002],
    'OrderDate': ['2024-01-05', '2024-01-06', '2024-01-07', '2024-01-10', '2024-01-12', '2024-01-15', '2024-01-18', '2024-01-20', '2024-01-22', '2024-01-25']
}
df_sales = pd.DataFrame(sales_data)

print(df_sales)                            # 모든 데이터 출력


```
1. 각 고객(`CustomerID`)별로 총 구매 금액을 계산하세요.
2. `Laptop`을 구매한 고객들의 평균 구매 금액을 계산하세요.
3. 가장 많이 판매된 상품(상품 이름과 판매된 총 수량)을 구하세요.
4. `OrderDate`가 2024년 1월 10일 이후인 주문만 필터링하여 출력하세요.
5. `Product`별로 판매량과 총 판매 금액을 계산한 후, 이를 `Price` 기준으로 내림차순으로 정렬하여 출력하세요.
```


#1. 각 고객(`CustomerID`)별로 총 구매 금액을 계산하세요.

df_sales['Total'] = df_sales['Quantity'] * df_sales['Price']

customer_total = df_sales.groupby('CustomerID')['Total'].sum()

print(customer_total)
